requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
--index-url http://predictail:jIXo005Bo+l7UCAfizoc@pypi.predictail.com/simple/
--extra-index-url http://pypi.python.org/simple
--trusted-host pypi.predictail.com
--trusted-host pypi.python.org

attrs==19.3.0             # via jsonschema
backcall==0.1.0           # via ipython
bleach==3.1.4             # via nbconvert
cachetools==4.1.0         # via google-auth
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
decorator==4.4.2          # via ipython, traitlets
defusedxml==0.6.0         # via nbconvert
entrypoints==0.3          # via nbconvert
fastavro==0.23.1
google-api-core[grpc]==1.16.0  # via google-cloud-bigquery, google-cloud-bigquery-storage, google-cloud-core
google-auth-oauthlib==0.4.1  # via pandas-gbq, pydata-google-auth
google-auth==1.13.1       # via google-api-core, google-auth-oauthlib, google-cloud-bigquery, pandas-gbq, pydata-google-auth
google-cloud-bigquery-storage==0.8.0
google-cloud-bigquery[pandas]==1.24.0
google-cloud-core==1.3.0  # via google-cloud-bigquery
google-resumable-media==0.5.0  # via google-cloud-bigquery
googleapis-common-protos==1.51.0  # via google-api-core
grpcio==1.28.1            # via google-api-core
idna==2.9                 # via requests
importlib-metadata==1.6.0  # via jsonschema
ipykernel==5.2.0          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via nbformat, notebook, qtconsole, traitlets
ipython==7.13.0
ipywidgets==7.5.1         # via jupyter
jedi==0.16.0              # via ipython
jinja2==2.11.1            # via nbconvert, notebook
jsonschema==3.2.0         # via nbformat
jupyter-client==6.1.2     # via ipykernel, jupyter-console, notebook, qtconsole
jupyter-console==6.1.0    # via jupyter
jupyter-core==4.6.3       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0
jupytext==1.4.2
markupsafe==1.1.1         # via jinja2
mistune==0.8.4            # via nbconvert
nbconvert==5.6.1          # via jupyter, notebook
nbformat==5.0.5           # via ipywidgets, jupytext, nbconvert, notebook
notebook==6.0.3           # via jupyter, widgetsnbextension
numpy==1.18.2             # via pandas, pyarrow
oauthlib==3.1.0           # via requests-oauthlib
pandas-gbq==0.13.1
pandas==1.0.3
pandocfilters==1.4.2      # via nbconvert
parso==0.6.2              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
prometheus-client==0.7.1  # via notebook
prompt-toolkit==3.0.5     # via ipython, jupyter-console
protobuf==3.11.3          # via google-api-core, google-cloud-bigquery, googleapis-common-protos
ptyprocess==0.6.0         # via pexpect, terminado
pyarrow==0.16.0
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pydata-google-auth==0.3.0  # via pandas-gbq
pygments==2.6.1           # via ipython, jupyter-console, nbconvert, qtconsole
pyrsistent==0.16.0        # via jsonschema
python-dateutil==2.8.1    # via jupyter-client, pandas
pytz==2019.3              # via fastavro, google-api-core, pandas
pyyaml==5.3.1             # via jupytext
pyzmq==19.0.0             # via jupyter-client, notebook, qtconsole
qtconsole==4.7.2          # via jupyter
qtpy==1.9.0               # via qtconsole
requests-oauthlib==1.3.0  # via google-auth-oauthlib
requests==2.23.0          # via google-api-core, requests-oauthlib
rsa==4.0                  # via google-auth
send2trash==1.5.0         # via notebook
six==1.14.0               # via bleach, google-api-core, google-auth, google-cloud-bigquery, google-resumable-media, grpcio, jsonschema, protobuf, pyarrow, pyrsistent, python-dateutil, traitlets
terminado==0.8.3          # via notebook
testpath==0.4.4           # via nbconvert
tornado==6.0.4            # via ipykernel, jupyter-client, notebook, terminado
tqdm==4.45.0
traitlets==4.3.3          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbconvert, nbformat, notebook, qtconsole
urllib3==1.25.8           # via requests
wcwidth==0.1.9            # via prompt-toolkit
webencodings==0.5.1       # via bleach
widgetsnbextension==3.5.1  # via ipywidgets
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==46.1.3        # via google-api-core, google-auth, ipython, jsonschema, pandas-gbq, protobuf, pydata-google-auth
